Re-apply TextBase.formattedText after a style property change
Resolves #1078
