+++ tsconfig.json
@@ -659,11 +659,13 @@
         "ui/tab-view/tab-view.android.ts",
         "ui/tab-view/tab-view.d.ts",
         "ui/tab-view/tab-view.ios.ts",
+        "ui/text-base/text-base-common.ts",
         "ui/text-base/text-base-styler.android.ts",
         "ui/text-base/text-base-styler.d.ts",
         "ui/text-base/text-base-styler.ios.ts",
+        "ui/text-base/text-base.android.ts",
         "ui/text-base/text-base.d.ts",
-        "ui/text-base/text-base.ts",
+        "ui/text-base/text-base.ios.ts",
         "ui/text-field/text-field-common.ts",
         "ui/text-field/text-field.android.ts",
         "ui/text-field/text-field.d.ts",
