+++ ui/button/button.android.ts
@@ -1,5 +1,6 @@
 ﻿import common = require("./button-common");
 import utils = require("utils/utils")
+import dependencyObservable = require("ui/core/dependency-observable");
 
 global.moduleMerge(common, exports);
 
@@ -35,5 +36,17 @@ export class Button extends common.Button {
                 }
             }
         }));
+    }
+
+    public _onTextPropertyChanged(data: dependencyObservable.PropertyChangeData) {
+        if (this.android) {
+            this.android.setText(data.newValue + "");
+        }
+    }
+
+    public _setFormattedTextPropertyToNative(value) {
+        if (this.android) {
+            this.android.setText(value._formattedText);
+        }
     }
 }
