+++ ui/text-base/text-base.android.ts
@@ -1,0 +1,17 @@
+﻿import common = require("./text-base-common");
+import types = require("utils/types");
+import dependencyObservable = require("ui/core/dependency-observable");
+
+export class TextBase extends common.TextBase {
+    public _onTextPropertyChanged(data: dependencyObservable.PropertyChangeData) {
+        if (this.android) {
+            var newValue = types.isNullOrUndefined(data.newValue) ? "" : data.newValue + "";
+            this.android.setText(newValue);
+        }
+    }
+    public _setFormattedTextPropertyToNative(value) {
+        if (this.android) {
+            this.android.setText(value._formattedText);
+        }
+    }
+}
